artifacts/mqtt-data-simulator/requirements.txt
paho-mqtt
colorama
pyfiglet